api/requirements.txt
Django==4.2
dnspython==2.0.0
PyMongo
django-ninja
python-dotenv
requests
bson